+++ cmake/BLTPrivateMacros.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2019, Lawrence Livermore National Security, LLC and
+# Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level COPYRIGHT file for details
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
@@ -92,25 +92,38 @@ endfunction()
 function(blt_fix_fortran_openmp_flags target_name)
 
     if (ENABLE_FORTRAN AND ENABLE_OPENMP AND BLT_OPENMP_FLAGS_DIFFER)
+        # The OpenMP interface library will have been added as a direct
+        # link dependency instead of via flags
+        get_target_property(target_link_libs ${target_name} LINK_LIBRARIES)
+        if ( target_link_libs )
+            # Since this is only called on executable targets we can safely convert
+            # from a "real" target back to a "fake" one as this is a sink vertex in
+            # the DAG.  Only the link flags need to be modified.
+            list(FIND target_link_libs "openmp" _omp_index)
+            if(${_omp_index} GREATER -1)
+                list(REMOVE_ITEM target_link_libs "openmp")
 
-        set(_property_name LINK_FLAGS)
-        if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
-            # In CMake 3.13+, LINK_FLAGS was converted to LINK_OPTIONS.
-            set(_property_name LINK_OPTIONS)
-        endif()
+                # Copy the compile flags verbatim
+                get_target_property(omp_compile_flags openmp INTERFACE_COMPILE_OPTIONS)
+                target_compile_options(${target_name} PUBLIC ${omp_compile_flags})
 
-        get_target_property(target_link_flags ${target_name} ${_property_name})
-        if ( target_link_flags )
+                # These are set through blt_add_target_link_flags which needs to use
+                # the link_libraries for interface libraries in CMake < 3.13
+                if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
+                    get_target_property(omp_link_flags openmp INTERFACE_LINK_OPTIONS)
+                else()
+                    get_target_property(omp_link_flags openmp INTERFACE_LINK_LIBRARIES)
+                endif()
+    
+                string( REPLACE "${OpenMP_CXX_FLAGS}" "${OpenMP_Fortran_FLAGS}"
+                        correct_omp_link_flags
+                        "${omp_link_flags}"
+                        )
+                list(APPEND target_link_libs "${correct_omp_link_flags}")
+                set_target_properties( ${target_name} PROPERTIES
+                                       LINK_LIBRARIES "${target_link_libs}" )
+            endif()
 
-            message(STATUS "Fixing OpenMP Flags for target[${target_name}]")
-
-            string( REPLACE "${OpenMP_CXX_FLAGS}" "${OpenMP_Fortran_FLAGS}"
-                    correct_link_flags
-                    "${target_link_flags}"
-                    )
-
-            set_target_properties( ${target_name} PROPERTIES ${_property_name}
-                                   "${correct_link_flags}" )
         endif()
 
     endif()
@@ -123,7 +136,8 @@ endfunction()
 ##
 ## This macro attempts to find the given executable via either a previously defined
 ## <UPPERCASE_NAME>_EXECUTABLE or using find_program with the given EXECUTABLES.
-## if EXECUTABLES is left empty, then NAME is used.
+## if EXECUTABLES is left empty, then NAME is used.  This macro will only attempt
+## to locate the executable if <UPPERCASE_NAME>_ENABLED is TRUE.
 ##
 ## If successful the following variables will be defined:
 ## <UPPERCASE_NAME>_FOUND
@@ -228,6 +242,20 @@ macro(blt_inherit_target_info)
         target_compile_definitions( ${arg_TO} PUBLIC ${_interface_defines})
     endif()
 
+    if( ${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.13.0" )
+        get_target_property(_interface_link_options
+                            ${arg_FROM} INTERFACE_LINK_OPTIONS)
+        if ( _interface_link_options )
+            target_link_options( ${arg_TO} PUBLIC ${_interface_link_options})
+        endif()
+    endif()
+
+    get_target_property(_interface_compile_options
+                        ${arg_FROM} INTERFACE_COMPILE_OPTIONS)
+    if ( _interface_compile_options )
+        target_compile_options( ${arg_TO} PUBLIC ${_interface_compile_options})
+    endif()
+
     if ( NOT arg_OBJECT )
         get_target_property(_interface_link_directories
                             ${arg_FROM} INTERFACE_LINK_DIRECTORIES)
@@ -243,6 +271,46 @@ macro(blt_inherit_target_info)
     endif()
 
 endmacro(blt_inherit_target_info)
+
+##------------------------------------------------------------------------------
+## blt_expand_depends( DEPENDS_ON [dep1 ...]
+##                     RESULT [variable] )
+##------------------------------------------------------------------------------
+macro(blt_expand_depends)
+    set(options)
+    set(singleValueArgs RESULT)
+    set(multiValueArgs DEPENDS_ON)
+
+    # Parse the arguments
+    cmake_parse_arguments(arg "${options}" "${singleValueArgs}"
+                        "${multiValueArgs}" ${ARGN} )
+
+    # Expand dependency list
+    set(_deps_to_process ${arg_DEPENDS_ON})
+    set(_expanded_DEPENDS_ON)
+    while(_deps_to_process)
+        # Copy the current set of dependencies to process
+        set(_current_deps_to_process ${_deps_to_process})
+        # and add them to the full expanded list
+        list(APPEND _expanded_DEPENDS_ON ${_deps_to_process})
+        # Then clear it so we can check if new ones were added
+        set(_deps_to_process)
+        foreach( dependency ${_current_deps_to_process} )
+            string(TOUPPER ${dependency} uppercase_dependency )
+            if ( DEFINED _BLT_${uppercase_dependency}_DEPENDS_ON )
+                foreach(new_dependency ${_BLT_${uppercase_dependency}_DEPENDS_ON})
+                    # Don't add duplicates
+                    if (NOT ${new_dependency} IN_LIST _expanded_DEPENDS_ON)
+                        list(APPEND _deps_to_process ${new_dependency})
+                    endif()
+                endforeach()
+            endif()
+        endforeach()
+    endwhile()
+
+    # Write the output to the requested variable
+    set(${arg_RESULT} ${_expanded_DEPENDS_ON})
+endmacro()
 
 
 ##------------------------------------------------------------------------------
@@ -265,42 +333,44 @@ macro(blt_setup_target)
         message( FATAL_ERROR "Must provide a NAME argument to the 'blt_setup_target' macro" )
     endif()
 
-    # Expand dependency list
-    set(_expanded_DEPENDS_ON ${arg_DEPENDS_ON})
-    foreach( i RANGE 50 )
-        foreach( dependency ${_expanded_DEPENDS_ON} )
-            string(TOUPPER ${dependency} uppercase_dependency )
+    # Default to "real" scope, unless it's an interface library
+    set(_private_scope PRIVATE)
+    set(_public_scope  PUBLIC)
+    get_target_property(_target_type ${arg_NAME} TYPE)
+    if("${_target_type}" STREQUAL "INTERFACE_LIBRARY")
+        set(_private_scope INTERFACE)
+        set(_public_scope  INTERFACE)
+    endif()
 
-            if ( DEFINED _BLT_${uppercase_dependency}_DEPENDS_ON )
-                foreach(new_dependency ${_BLT_${uppercase_dependency}_DEPENDS_ON})
-                    if (NOT ${new_dependency} IN_LIST _expanded_DEPENDS_ON)
-                        list(APPEND _expanded_DEPENDS_ON ${new_dependency})
-                    endif()
-                endforeach()
-            endif()
-        endforeach()
-    endforeach()
+    # Expand dependency list - avoid "recalculating" if the information already exists
+    set(_expanded_DEPENDS_ON)
+    if(NOT "${_target_type}" STREQUAL "INTERFACE_LIBRARY")
+        get_target_property(_expanded_DEPENDS_ON ${arg_NAME} BLT_EXPANDED_DEPENDENCIES)
+    endif()
+    if(NOT _expanded_DEPENDS_ON)
+        blt_expand_depends(DEPENDS_ON ${arg_DEPENDS_ON} RESULT _expanded_DEPENDS_ON)
+    endif()
 
     # Add dependency's information
     foreach( dependency ${_expanded_DEPENDS_ON} )
         string(TOUPPER ${dependency} uppercase_dependency )
 
         if ( NOT arg_OBJECT AND _BLT_${uppercase_dependency}_IS_OBJECT_LIBRARY )
-            target_sources(${arg_NAME} PRIVATE $<TARGET_OBJECTS:${dependency}>)
+            target_sources(${arg_NAME} ${_private_scope} $<TARGET_OBJECTS:${dependency}>)
         endif()
 
         if ( DEFINED _BLT_${uppercase_dependency}_INCLUDES )
             if ( _BLT_${uppercase_dependency}_TREAT_INCLUDES_AS_SYSTEM )
-                target_include_directories( ${arg_NAME} SYSTEM PUBLIC
+                target_include_directories( ${arg_NAME} SYSTEM ${_public_scope}
                     ${_BLT_${uppercase_dependency}_INCLUDES} )
             else()
-                target_include_directories( ${arg_NAME} PUBLIC
+                target_include_directories( ${arg_NAME} ${_public_scope}
                     ${_BLT_${uppercase_dependency}_INCLUDES} )
             endif()
         endif()
 
         if ( DEFINED _BLT_${uppercase_dependency}_FORTRAN_MODULES )
-            target_include_directories( ${arg_NAME} PUBLIC
+            target_include_directories( ${arg_NAME} ${_public_scope}
                 ${_BLT_${uppercase_dependency}_FORTRAN_MODULES} )
         endif()
 
@@ -331,15 +401,15 @@ macro(blt_setup_target)
             # actual CMake targets
             if(NOT "${_BLT_${uppercase_dependency}_LIBRARIES}"
                     STREQUAL "BLT_NO_LIBRARIES" )
-                target_link_libraries( ${arg_NAME} PUBLIC
+                target_link_libraries( ${arg_NAME} ${_public_scope}
                     ${_BLT_${uppercase_dependency}_LIBRARIES} )
             endif()
         else()
-            target_link_libraries( ${arg_NAME} PUBLIC ${dependency} )
+            target_link_libraries( ${arg_NAME} ${_public_scope} ${dependency} )
         endif()
 
         if ( DEFINED _BLT_${uppercase_dependency}_DEFINES )
-            target_compile_definitions( ${arg_NAME} PUBLIC
+            target_compile_definitions( ${arg_NAME} ${_public_scope}
                 ${_BLT_${uppercase_dependency}_DEFINES} )
         endif()
 
@@ -352,7 +422,18 @@ macro(blt_setup_target)
             blt_add_target_link_flags(TO ${arg_NAME}
                                       FLAGS ${_BLT_${uppercase_dependency}_LINK_FLAGS} )
         endif()
-
+        # Propagate the overridden linker language, if applicable
+        if(TARGET ${dependency})
+            # If it's an interface library CMake doesn't even allow us to query the property
+            get_target_property(_dep_type ${dependency} TYPE)
+            if(NOT "${_dep_type}" STREQUAL "INTERFACE_LIBRARY")
+                get_target_property(_blt_link_lang ${dependency} INTERFACE_BLT_LINKER_LANGUAGE_OVERRIDE)
+                # TODO: Do we need to worry about overwriting?  Should only ever be HIP or CUDA
+                if(_blt_link_lang)
+                    set_target_properties(${arg_NAME} PROPERTIES INTERFACE_BLT_LINKER_LANGUAGE_OVERRIDE ${_blt_link_lang})
+                endif()
+            endif()
+        endif()
     endforeach()
 
 endmacro(blt_setup_target)
@@ -398,6 +479,9 @@ macro(blt_setup_cuda_target)
     if (${_depends_on_cuda_runtime} OR ${_depends_on_cuda})
         if (CUDA_LINK_WITH_NVCC)
             set_target_properties( ${arg_NAME} PROPERTIES LINKER_LANGUAGE CUDA)
+            # This will be propagated up to executable targets that depend on this
+            # library, which will need the HIP linker
+            set_target_properties( ${arg_NAME} PROPERTIES INTERFACE_BLT_LINKER_LANGUAGE_OVERRIDE CUDA)
         endif()
     endif()
 
@@ -432,6 +516,38 @@ macro(blt_setup_cuda_target)
         endif()
     endif()
 endmacro(blt_setup_cuda_target)
+
+##------------------------------------------------------------------------------
+## blt_cleanup_hip_globals(FROM_TARGET <target>)
+##
+## Needed as the SetupHIP macros (specifically, HIP_PREPARE_TARGET_COMMANDS)
+## "pollutes" the global HIP_HIPCC_FLAGS with target-specific options.  This
+## macro removes the target-specific generator expressions from the global flags
+## which have already been copied to source-file-specific instances of the
+## run_hipcc script.  Other global flags in HIP_HIPCC_FLAGS, e.g., those set by
+## the user, are left untouched.
+##------------------------------------------------------------------------------
+macro(blt_cleanup_hip_globals)
+    set(options)
+    set(singleValueArgs FROM_TARGET)
+    set(multiValueArgs)
+
+    # Parse the arguments
+    cmake_parse_arguments(arg "${options}" "${singleValueArgs}"
+                            "${multiValueArgs}" ${ARGN} )
+
+    # Check arguments
+    if ( NOT DEFINED arg_FROM_TARGET )
+        message( FATAL_ERROR "Must provide a FROM_TARGET argument to the 'blt_cleanup_hip_globals' macro")
+    endif()
+
+    # Remove the compile definitions generator expression
+    # This must be copied verbatim from HIP_PREPARE_TARGET_COMMANDS,
+    # which would have just added it to HIP_HIPCC_FLAGS
+    set(_defines_genexpr "$<TARGET_PROPERTY:${arg_FROM_TARGET},COMPILE_DEFINITIONS>")
+    set(_defines_flags_genexpr "$<$<BOOL:${_defines_genexpr}>:-D$<JOIN:${_defines_genexpr}, -D>>")
+    list(REMOVE_ITEM HIP_HIPCC_FLAGS ${_defines_flags_genexpr})
+endmacro(blt_cleanup_hip_globals)
 
 ##------------------------------------------------------------------------------
 ## blt_add_hip_library(NAME         <libname>
@@ -486,8 +602,18 @@ macro(blt_add_hip_library)
                                      HIP_SOURCE_PROPERTY_FORMAT TRUE)
 
         hip_add_library( ${arg_NAME} ${arg_SOURCES} ${arg_LIBRARY_TYPE} )
+        blt_cleanup_hip_globals(FROM_TARGET ${arg_NAME})
+        # Link to the hip_runtime target so it gets pulled in by targets
+        # depending on this target
+        target_link_libraries(${arg_NAME} PUBLIC hip_runtime)
     else()
         add_library( ${arg_NAME} ${arg_LIBRARY_TYPE} ${arg_SOURCES} ${arg_HEADERS} )
+    endif()
+
+    if (${_depends_on_hip_runtime} OR ${_depends_on_hip})
+        # This will be propagated up to executable targets that depend on this
+        # library, which will need the HIP linker
+        set_target_properties( ${arg_NAME} PROPERTIES INTERFACE_BLT_LINKER_LANGUAGE_OVERRIDE HIP)
     endif()
 
 endmacro(blt_add_hip_library)
@@ -529,6 +655,20 @@ macro(blt_add_hip_executable)
         set(_depends_on_hip_runtime TRUE)
     endif()
 
+    blt_expand_depends(DEPENDS_ON ${arg_DEPENDS_ON} RESULT _expanded_DEPENDS_ON)
+    foreach( dependency ${_expanded_DEPENDS_ON} )
+        if(TARGET ${dependency})
+            get_target_property(_dep_type ${dependency} TYPE)
+            if(NOT "${_dep_type}" STREQUAL "INTERFACE_LIBRARY")
+                # Propagate the overridden linker language, if applicable
+                get_target_property(_blt_link_lang ${dependency} INTERFACE_BLT_LINKER_LANGUAGE_OVERRIDE)
+                if(_blt_link_lang STREQUAL "HIP")
+                    set(_depends_on_hip_runtime TRUE)
+                endif()
+            endif()
+        endif()
+    endforeach()
+
     if (${_depends_on_hip} OR ${_depends_on_hip_runtime})
         # if hip is in depends_on, flag each file's language as HIP
         # instead of leaving it up to CMake to decide
@@ -544,26 +684,35 @@ macro(blt_add_hip_executable)
                                      HIP_SOURCE_PROPERTY_FORMAT TRUE)
 
         hip_add_executable( ${arg_NAME} ${arg_SOURCES} )
+        blt_cleanup_hip_globals(FROM_TARGET ${arg_NAME})
     else()
         add_executable( ${arg_NAME} ${arg_SOURCES} ${arg_HEADERS})
     endif()
+
+    # Save the expanded dependencies to avoid recalculating later
+    set_target_properties(${arg_NAME} PROPERTIES
+                          BLT_EXPANDED_DEPENDENCIES "${_expanded_DEPENDS_ON}")
 
 endmacro(blt_add_hip_executable)
 
 ##------------------------------------------------------------------------------
 ## blt_split_source_list_by_language( SOURCES <sources>
 ##                                    C_LIST <list name>
-##                                    Fortran_LIST <list name>)
+##                                    Fortran_LIST <list name>
+##                                    Python_LIST <list name>)
+##                                    CMAKE_LIST <list name>)
 ##
-## Filters source list by file extension into C/C++ and Fortran source lists
-## based on BLT_C_FILE_EXTS and BLT_Fortran_FILE_EXTS (global BLT variables).
-## Files with no extension or generator expressions that are not object libraries
-## (of the form "$<TARGET_OBJECTS:nameofobjectlibrary>") will throw fatal errors.
-##------------------------------------------------------------------------------
+## Filters source list by file extension into C/C++, Fortran, Python, and
+## CMake source lists based on BLT_C_FILE_EXTS, BLT_Fortran_FILE_EXTS,
+## and BLT_CMAKE_FILE_EXTS (global BLT variables). Files named
+## "CMakeLists.txt" are also filtered here. Files with no extension
+## or generator expressions that are not object libraries (of the form
+## "$<TARGET_OBJECTS:nameofobjectlibrary>") will throw fatal errors.
+## ------------------------------------------------------------------------------
 macro(blt_split_source_list_by_language)
 
     set(options)
-    set(singleValueArgs C_LIST Fortran_LIST)
+    set(singleValueArgs C_LIST Fortran_LIST Python_LIST CMAKE_LIST)
     set(multiValueArgs SOURCES)
 
     # Parse the arguments
@@ -585,23 +734,33 @@ macro(blt_split_source_list_by_language)
             message(FATAL_ERROR "blt_split_source_list_by_language macro does not support generator expressions because CMake does not provide a way to evaluate them. Given generator expression: ${_file}")
         endif()
 
-        get_filename_component(_ext ${_file} EXT)
+        get_filename_component(_ext "${_file}" EXT)
         if("${_ext}" STREQUAL "")
             message(FATAL_ERROR "blt_split_source_list_by_language given source file with no extension: ${_file}")
         endif()
 
-        string(TOLOWER ${_ext} _ext_lower)
+        get_filename_component(_name "${_file}" NAME)  
 
-        if(${_ext_lower} IN_LIST BLT_C_FILE_EXTS)
+        string(TOLOWER "${_ext}" _ext_lower)
+
+        if("${_ext_lower}" IN_LIST BLT_C_FILE_EXTS)
             if (DEFINED arg_C_LIST)
-                list(APPEND ${arg_C_LIST} ${_file})
+                list(APPEND ${arg_C_LIST} "${_file}")
             endif()
-        elseif(${_ext_lower} IN_LIST BLT_Fortran_FILE_EXTS)
+        elseif("${_ext_lower}" IN_LIST BLT_Fortran_FILE_EXTS)
             if (DEFINED arg_Fortran_LIST)
-                list(APPEND ${arg_Fortran_LIST} ${_file})
+                list(APPEND ${arg_Fortran_LIST} "${_file}")
+            endif()
+        elseif("${_ext_lower}" IN_LIST BLT_Python_FILE_EXTS)
+            if (DEFINED arg_Python_LIST)
+                list(APPEND ${arg_Python_LIST} "${_file}")
+            endif()
+        elseif("${_ext_lower}" IN_LIST BLT_CMAKE_FILE_EXTS OR "${_name}" STREQUAL "CMakeLists.txt")
+            if (DEFINED arg_CMAKE_LIST)
+                list(APPEND ${arg_CMAKE_LIST} "${_file}")
             endif()
         else()
-            message(FATAL_ERROR "blt_split_source_list_by_language given source file with unknown file extension. Add the missing extension to the corresponding list (BLT_C_FILE_EXTS or BLT_Fortran_FILE_EXTS).\n Unknown file: ${_file}")
+            message(FATAL_ERROR "blt_split_source_list_by_language given source file with unknown file extension. Add the missing extension to the corresponding list (BLT_C_FILE_EXTS, BLT_Fortran_FILE_EXTS, BLT_Python_FILE_EXTS, or BLT_CMAKE_FILE_EXTS).\n Unknown file: ${_file}")
         endif()
     endforeach()