+++ cmake/thirdparty/SetupHIP.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-2019, Lawrence Livermore National Security, LLC and
+# Copyright (c) 2017-2021, Lawrence Livermore National Security, LLC and
 # other BLT Project Developers. See the top-level COPYRIGHT file for details
 #
 # SPDX-License-Identifier: (BSD-3-Clause)
@@ -14,8 +14,6 @@ find_package(HIP REQUIRED)
 
 message(STATUS "HIP version:      ${HIP_VERSION_STRING}")
 message(STATUS "HIP platform:     ${HIP_PLATFORM}")
-#message(STATUS "HIP Include Path: ${HIP_INCLUDE_DIRS}")
-#message(STATUS "HIP Libraries:    ${HIP_LIBRARIES}")
 
 set(HIP_RUNTIME_INCLUDE_DIRS "${HIP_ROOT_DIR}/include")
 if(${HIP_PLATFORM} STREQUAL "hcc")
@@ -30,63 +28,31 @@ elseif(${HIP_PLATFORM} STREQUAL "nvcc")
     set(HIP_RUNTIME_LIBRARIES "${CUDAToolkit_LIBRARY_DIR}/libcudart.so")
     set(HIP_RUNTIME_INCLUDE_DIRS "${HIP_RUNTIME_INCLUDE_DIRS};${CUDAToolkit_INCLUDE_DIR}")
 endif()
-set(HIP_RUNTIME_INCLUDE_DIRS "${HIP_ROOT_DIR}/include;${HIP_ROOT_DIR}/hcc/include" CACHE STRING "")
-set(HIP_RUNTIME_COMPILE_FLAGS "${HIP_RUNTIME_COMPILE_FLAGS};${HIP_RUNTIME_DEFINES}")
-# set(HIP_RUNTIME_LIBRARIES "${HIP_ROOT_DIR}/hcc/lib")
-# set(HIP_RUNTIME_LIBRARIES "${HIP_ROOT_DIR}/hcc/lib")
-
-set(_hip_compile_flags " ")
-if (ENABLE_CLANG_HIP)
-    set(_hip_compile_flags -x hip)
-    # Using clang HIP, we need to construct a few CPP defines and compiler flags
-    foreach(arch ${BLT_CLANG_HIP_ARCH})
-        string(TOUPPER ${arch} UPARCH)
-        string(TOLOWER ${arch} lowarch)
-        list(APPEND _hip_compile_flags "--offload-arch=${lowarch}" "-D__HIP_ARCH_${UPARCH}__=1")
-    endforeach(arch)
-    
-    # We need to pass rocm path as well, for certain bitcode libraries.
-    # First see if we were given it, then see if it exists in the environment.
-    # If not, don't try to guess but print a warning and hope the compiler knows where it is.
-    if(NOT ROCM_PATH)
-        if(DEFINED ENV{ROCM_PATH})
-            set(ROCM_PATH "$ENV{ROCM_PATH}")
-        endif()
-    endif()
-
-    if(DEFINED ROCM_PATH)
-        list(APPEND _hip_compile_flags "--rocm-path=${ROCM_PATH}")
-    else()
-        message(WARNING "ROCM_PATH not found. Set this if the compiler can't find device bitcode libraries.")
-    endif()
-
-    message(STATUS "Clang HIP Enabled. HIP compile flags added: ${_hip_compile_flags}")
-
-    # Fundamendally this is just the runtime includes with some extra compile flags
-    blt_register_library(NAME      hip
-                         COMPILE_FLAGS ${_hip_compile_flags}
-                         INCLUDES  ${HIP_RUNTIME_INCLUDE_DIRS}
-                         LIBRARIES ${HIP_RUNTIME_LIBRARIES}
-                         TREAT_INCLUDES_AS_SYSTEM ON)
-
+if ( IS_DIRECTORY "${HIP_ROOT_DIR}/hcc/include" ) # this path only exists on older rocm installs
+        set(HIP_RUNTIME_INCLUDE_DIRS "${HIP_ROOT_DIR}/include;${HIP_ROOT_DIR}/hcc/include" CACHE STRING "")
 else()
-
-    # depend on 'hip', if you need to use hip
-    # headers, link to hip libs, and need to run your source
-    # through a hip compiler (hipcc)
-    blt_register_library(NAME      hip
-                        INCLUDES  ${HIP_INCLUDE_DIRS}
-                        LIBRARIES ${HIP_LIBRARIES}
-                        TREAT_INCLUDES_AS_SYSTEM ON)
+        set(HIP_RUNTIME_INCLUDE_DIRS "${HIP_ROOT_DIR}/include" CACHE STRING "")
 endif()
+set(HIP_RUNTIME_COMPILE_FLAGS "${HIP_RUNTIME_COMPILE_FLAGS};-D${HIP_RUNTIME_DEFINE};-Wno-unused-parameter")
 
+# depend on 'hip', if you need to use hip
+# headers, link to hip libs, and need to run your source
+# through a hip compiler (hipcc)
+# This is currently used only as an indicator for blt_add_hip* -- FindHIP/hipcc will handle resolution
+# of all required HIP-related includes/libraries/flags.
+if (ENABLE_CLANG_HIP)
+    blt_import_library(NAME      hip,
+                       COMPILE_FLAGS "-x hip")
+else
+    blt_import_library(NAME      hip)
+endif()
 
 # depend on 'hip_runtime', if you only need to use hip
 # headers or link to hip libs, but don't need to run your source
 # through a hip compiler (hipcc)
-blt_register_library(NAME          hip_runtime
-                     INCLUDES      ${HIP_RUNTIME_INCLUDE_DIRS}
-                     DEFINES       ${HIP_RUNTIME_DEFINES}
-                     COMPILE_FLAGS ${HIP_RUNTIME_COMPILE_FLAGS}
-                     LIBRARIES     ${HIP_RUNTIME_LIBRARIES}
-                     TREAT_INCLUDES_AS_SYSTEM ON)
+blt_import_library(NAME          hip_runtime
+                   INCLUDES      ${HIP_RUNTIME_INCLUDE_DIRS}
+                   DEFINES       ${HIP_RUNTIME_DEFINES}
+                   COMPILE_FLAGS ${HIP_RUNTIME_COMPILE_FLAGS}
+                   TREAT_INCLUDES_AS_SYSTEM ON
+                   EXPORTABLE    ${BLT_EXPORT_THIRDPARTY})